requirements.txt
requests
pydantic-ai
ddgs
python-dotenv